meta-lxatac-software: bcu: enable building bcu-native for use on host
This is not really relevant for the LXA TAC or meta-lxatac in general,
because the LXA TAC is not an i.MX EVK board.
But it makes sense in the context of upstreaming the recipe to some other
layer, where users may want to use `oe-run-native bcu-native bcu` to
control an EVK board.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>