meta-lxatac-bsp: linux-lxatac: add aliases for can/i2c/spi and disable rtc
The aliases ensure a stable numbering of the can / i2c / spi devices.
This solves an issue we have started to observe with Linux 6.11,
where the SPI bus numbers have changed/are no longer stable between boots.

The other change is disabling the RTC node, since it could not be used
anyways because the RTC is missing a 32.768 kHz clock crystal.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>