meta-lxatac-software: podman: add patch to start in multi-user.target
By default the podman.service starts as part of the default.target.
This means it also starts on the very first boot of a newly LXA TAC
image, where we boot into `system-update.target` instead of
`multi-user.target`.

We do not want that, because it ends up cluttering `/srv/containers`
on the root filesystem with files, instead of writing to the partition
that will be mounted there on later boots.

Fix that by applying a patch for now, while upstreaming it at the
same time for later.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>