meta-lxatac-software: distro: tacos: reduce permissions on rootfs /srv
We do not want any service to write to /srv (in the root filesystem)
before the /srv partition is set up and mounted.

Indicate this intent by reducing permissions for the directory.

This will however not stop processes running as root from writing there
anyways due to CAP_DAC_OVERRIDE.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>