meta-lxatac-bsp: lxatac-net-switch-hacks: increase atomic mem pool to 8M
When experiencing load the kernels default min_free_kbytes (of around 2M)
seem to little. Hot paths can run out of memory.
Increasing the limit to 8M seems to mitigate the problem.

This manifests in issues in communicating with the ethernet switch under
high loads, resulting in network connection losses.

This is only fighting symptoms of an underlying issue,
which why it is marked as a hack.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>