meta-lxatac-bsp: lxatac-net-switch-hacks: increase prio for SPI irq thread
When the system is under high load some SPI transfers with the ethernet
switch will time out before they are handled.

Increase the priority of the kernel thread that handles the SPI transfer
to work around the issue.

It does not make a lot of sense for a SPI transfer, that is 100% under the
hosts control (it does not and can not wait for the device for example) to
time out in the first place.
This means we are only fighting symptoms here, which is why this change
is also marked as a hack.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>