meta-lxatac-software: lxatac-core-image-base: install networkmanager meta package
We define DISTRO_FEATURES such as "wifi", "bluetooth" or "systemd"
already. Let's make use of that and install the networkmanager meta
package. More useful features are pulled in depending on the
DISTRO_FEATURES and PACKAGECONFIG options set.

  $ bitbake-getvar -r networkmanager PACKAGECONFIG
  [...]
  PACKAGECONFIG="readline nss ifupdown dnsmasq nmcli vala systemd bluez5 polkit wifi nmtui"

nmcli is still installed due to it being a part of networkmanager's
default PACKAGECONFIG.
nmtui is also installed due to our networkmanager_%.bbappend appending
it to networkmanager's PACKAGECONFIG.

With this change, wifi and bluetooth functionality is now available.

Signed-off-by: Bastian Krause <bst@pengutronix.de>